backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
SQLAlchemy>=2.0.31
chromadb>=0.5.3
langchain>=0.2.6
langchain-openai>=0.1.9
langchain-community>=0.2.6
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.2
langchain-text-splitters>=0.2.0
python-multipart>=0.0.9
langchain-google-genai>=1.0.6
pymongo>=4.8.0
langchain-mongodb>=0.1.6
dnspython>=2.6.1
